[build] Optimize imports

+++ structures/src/main/java/fr/laas/fape/structures/IR2IntMap.java
@@ -70,7 +70,7 @@ public class IR2IntMap<K> implements Map<K, Integer> {
     public Integer get(Object o) {
         if (!(o instanceof Identifiable)) throw new InvalidParameterException("Parameter is not identifiable: " + o);
         if (!containsKey(o)) throw new NoSuchElementException();
-        return (Integer) values[keyRep.asInt((K) o)];
+        return values[keyRep.asInt((K) o)];
     }
 
     @SuppressWarnings("unchecked")
@@ -117,7 +117,7 @@ public class IR2IntMap<K> implements Map<K, Integer> {
     }
 
     @Override public Integer putIfAbsent(K k, Integer v) {
-        if (v == null || NIL == ((int) v)) throw new InvalidParameterException("This map uses "+NIL+" to represent the absence of value");
+        if (v == null || NIL == v) throw new InvalidParameterException("This map uses "+NIL+" to represent the absence of value");
         if(!containsKey(k))
             return put(k, v);
         else
@@ -129,7 +129,7 @@ public class IR2IntMap<K> implements Map<K, Integer> {
     public Integer remove(Object o) {
         if (!(o instanceof Identifiable)) throw new InvalidParameterException("Parameter is not identifiable: " + o);
         if (!containsKey(o)) return null; //throw new NoSuchElementException();
-        Integer val = (Integer) values[keyRep.asInt((K) o)];
+        Integer val = values[keyRep.asInt((K) o)];
         values[keyRep.asInt((K) o)] = NIL;
         numElem--;
         return val;
@@ -208,7 +208,7 @@ public class IR2IntMap<K> implements Map<K, Integer> {
         List<Integer> vals = new ArrayList<>();
         for (int val : values)
             if (val != NIL)
-                vals.add((Integer) val);
+                vals.add(val);
         return vals;
     }
